emotion-relationship-recognition/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
mtcnn>=0.1.1
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.28.0
transformers>=4.30.0
sentencepiece>=0.1.99
albumentations>=1.3.0
tqdm>=4.65.0
tensorflow>=2.13.0
